Extract array with a member type into Array::Member
