+++ spec/dry/types/array_spec.rb
@@ -29,9 +29,10 @@ RSpec.describe Dry::Types::Array do
         end
 
         it 'raises when input is not valid' do
-          expect {
-            array[%w(1 2 3)]
-          }.to raise_error(Dry::Types::ConstraintError)
+          expect { array[%w(1 2 3)] }.to raise_error(
+            Dry::Types::ConstraintError,
+            '1 violates constraints (gt?(2) failed)'
+          )
         end
       end
     end
